Rename to plot, add deprecation warnings, remove typing
